+++ tools/make-zip.js
@@ -17,6 +17,8 @@ const include = [
   'manifest.json',
   'popup.html',
   'popup.js',
+  'options.html',
+  'options.js',
   'background.js',
   'content',
   'icons/icon16.png',